+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- `ADD` Administrative areas can be imported
 - `FIX` The Kafka host can be configured via an env var
 - `ADD` Published message details can be viewed
 - `ADD` Published messages are shown in the all messages list